Change Parquet API interaction to use `u64` (support files larger than 4GB in WASM) (#7371)
* Change AsyncFileReader trait for u64


* update metadatafetch trait


* Fix lint issue

* fix tests for latest main

* Address comments by @mbrobbel from #7252

* Fix compile

* Revert suffix length back to usize

* Use `u64` for `file_size`

* address comments

* fix calculation of metadata_start

* change file_size type to u64

* change _sized functions to take u64 file_size

* clippy

* remove some potential panics

* use u64 for page index ranges

* Revert change to deprecated method

---------

+++ parquet/tests/arrow_reader/bad_data.rs
@@ -156,7 +156,7 @@ async fn bad_metadata_err() {
 
     let metadata_buffer = Bytes::from_static(include_bytes!("bad_raw_metadata.bin"));
 
-    let metadata_length = metadata_buffer.len();
+    let metadata_length = metadata_buffer.len() as u64;
 
     let mut reader = std::io::Cursor::new(&metadata_buffer);
     let mut loader = ParquetMetaDataReader::new();
